Update help to match environment when running in wash shell
When running in the `wash` shell, we alias several commands so they
don't override common shell builtins. However the `help` text wasn't
updated to reflect this, which can lead to unexpected errors like trying
to run `exec` and having it overwrite the shell process rather than
invoke a command on the target.

Update help output for aliased commands when run from within the shell
to refer to only command names that are setup in the shell.

Fixes #329.

Signed-off-by: Michael Smith <michael.smith@puppet.com>